requirements.txt
aiogram>=3.14,<4.0
stripe>=10.0,<12.0
aiohttp>=3.10,<4.0
aiosqlite>=0.20,<1.0
python-dotenv>=1.0,<2.0
apscheduler>=3.10,<4.0
